+++ src/type.hpp
@@ -1,9 +1,11 @@
 #ifndef TYPE_HPP
 #define TYPE_HPP
 
+#include <iostream>
+
 #include "./ast.hpp"
 
-enum TypeSpec { VOID, I64, I32, BOOL, UNKNOWN, ERROR, POINTER, I8, VARIATIC, U64, U32, U8, F32, F64 };
+enum TypeSpec { VOID, I64, I32, BOOL, UNKNOWN, ERROR, POINTER, I8, VARIATIC, U64, U32, U8, F32, F64, ARRAY };
 
 auto operator<<(std::ostream& os, TypeSpec const& ts) -> std::ostream&;
 
@@ -87,6 +89,10 @@ class Type {
         return t_ == TypeSpec::BOOL;
     }
 
+    auto is_array() const noexcept -> bool {
+        return t_ == TypeSpec::ARRAY;
+    }
+
  protected:
     TypeSpec t_;
 };
@@ -127,6 +133,59 @@ class PointerType : public Type {
 
  private:
     std::shared_ptr<Type> sub_type_;
+};
+
+class ArrayType : public Type {
+ public:
+    ArrayType(std::shared_ptr<Type> sub_type)
+    : Type(TypeSpec::ARRAY)
+    , sub_type_(sub_type) {}
+
+    ArrayType(std::shared_ptr<Type> sub_type, size_t length)
+    : Type(TypeSpec::ARRAY)
+    , sub_type_(sub_type)
+    , length_(std::optional(length)) {}
+
+    auto get_sub_type() const -> std::shared_ptr<Type> {
+        return sub_type_;
+    }
+
+    auto get_length() const -> std::optional<size_t> {
+        return length_;
+    }
+
+    auto print(std::ostream& os) const -> void override {
+        os << t_;
+        sub_type_->print(os);
+    }
+
+    auto equals(const Type& other) const -> bool override {
+        if (t_ != other.get_type_spec())
+            return false;
+
+        auto* other_ptr = dynamic_cast<const ArrayType*>(&other);
+        if (!other_ptr)
+            return false;
+
+        if (length_.has_value() and other_ptr->get_length().has_value()) {
+            return *sub_type_ == *other_ptr->sub_type_ and *length_ == *other_ptr->get_length();
+        }
+        else {
+            return *sub_type_ == *other_ptr->sub_type_;
+        }
+    }
+
+    auto equal_soft(const Type& other) const -> bool override {
+        auto* other_ptr = dynamic_cast<const ArrayType*>(&other);
+        if (!other_ptr) {
+            return false;
+        }
+        return soft_typespec_equals(t_, other.get_type_spec()) and sub_type_->equal_soft(*other_ptr->get_sub_type());
+    }
+
+ private:
+    std::shared_ptr<Type> sub_type_;
+    std::optional<size_t> length_ = std::nullopt;
 };
 
 auto operator<<(std::ostream& os, Type const& t) -> std::ostream&;