+++ src/visitor.hpp
@@ -20,6 +20,7 @@ class CharExpr;
 class VarExpr;
 class CallExpr;
 class CastExpr;
+class ArrayInitExpr;
 class EmptyStmt;
 class CompoundStmt;
 class LocalVarStmt;
@@ -50,6 +51,7 @@ class Visitor {
     virtual auto visit_var_expr(std::shared_ptr<VarExpr> var_expr) -> void = 0;
     virtual auto visit_call_expr(std::shared_ptr<CallExpr> call_expr) -> void = 0;
     virtual auto visit_cast_expr(std::shared_ptr<CastExpr> cast_expr) -> void = 0;
+    virtual auto visit_array_init_expr(std::shared_ptr<ArrayInitExpr> array_init_expr) -> void = 0;
 
     virtual auto visit_empty_stmt(std::shared_ptr<EmptyStmt> empty_stmt) -> void = 0;
     virtual auto visit_compound_stmt(std::shared_ptr<CompoundStmt> compound_stmt) -> void = 0;